refactor(tests): reorganize test files and mocks into unified tests directory
Move all test-related files from config/tests to a new tests directory at src/tests for better organization
Add new mock files and test utilities to support testing
Update all test file imports to reference new location
