+++ src/domain/auth/dtos/__tests__/login.dto.test.ts
@@ -1,5 +1,5 @@
 import { LoginDto } from "../login.dto";
-import { clearAllMocks } from "@/config/tests/test-utils";
+import { clearAllMocks } from "@/tests/test-utils";
 
 describe("LoginDto", () => {
   afterEach(() => {
