+++ src/domain/auth/dtos/__tests__/request-reset-password-email.dto.test.ts
@@ -1,5 +1,5 @@
 import { RequestResetPasswordEmailDto } from "../request-reset-password-email.dto";
-import { clearAllMocks } from "@/config/tests/test-utils";
+import { clearAllMocks } from "@/tests/test-utils";
 
 describe("RequestResetPasswordEmailDto", () => {
   afterEach(() => {
