+++ src/domain/auth/use-cases/__test__/request-reset-password-email.test.ts
@@ -1,6 +1,6 @@
 import { RequestResetPasswordEmail } from "../request-reset-password-email";
 import { RequestResetPasswordEmailDto } from "../../dtos/request-reset-password-email.dto";
-import { MockAuthRepository } from "../../../../config/tests/__mocks__/auth-repository.mock";
+import { MockAuthRepository } from "@/tests/__mocks__/auth-repository.mock";
 
 describe("RequestResetPasswordEmail", () => {
   let requestResetPasswordEmail: RequestResetPasswordEmail;
