+++ src/domain/user/use-cases/__test__/update-user.test.ts
@@ -1,7 +1,7 @@
 import { UpdateUser } from "@/domain/user/use-cases/update-user";
 import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
 import { UserEntity } from "@/domain/user/entities/user.entity";
-import { MockAuthRepository } from "@/config/tests/__mocks__/auth-repository.mock";
+import { MockAuthRepository } from "@/tests/__mocks__/auth-repository.mock";
 
 describe("UpdateUser", () => {
   let updateUser: UpdateUser;
