+++ src/infrastructure/external/auth/datasources/__tests__/set-session.test.ts
@@ -12,7 +12,7 @@ import {
   createMockDatasourceUserMapper,
   createMockUserEntity,
   createMockAuthUserEntity,
-} from "@/config/tests/__helpers__/auth-datasource.helpers";
+} from "@/tests/__helpers__/auth-datasource.helpers";
 
 // Mock dependencies
 jest.mock("@/infrastructure/external/auth/auth.client");
